# go/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "统一编码格式-UTF8\n",
        "title": "验证系统 API",
        "contact": {},
        "license": {
            "name": "易轩软件 http://www.99v66.com",
            "url": "http://www.99v66.com"
        },
        "version": "1.0"
    },
    "host": "127.0.0.1:52777",
    "basePath": "/v1",
    "paths": {
        "/apply/add": {
            "post": {
                "description": "详细描述",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "应用管理"
                ],
                "summary": "添加应用",
                "parameters": [
                    {
                        "type": "string",
                        "description": "登录返回的token",
                        "name": "token",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "应用名称",
                        "name": "name",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "收费模式 1=扣时 2=扣点",
                        "name": "mold",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "绑定模式 1=不绑定 2=绑定机器码 3=绑定IP",
                        "name": "bin",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "登录模式 1=免费 2=单点登录 3=多点登录",
                        "name": "login",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "远程数据",
                        "name": "msg",
                        "in": "formData"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"code\":200,\"msg\":\"添加成功\"}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/apply/edit": {
            "post": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "应用管理"
                ],
                "summary": "编辑应用信息",
                "parameters": [
                    {
                        "type": "string",
                        "description": "登录返回的token",
                        "name": "token",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "应用ID",
                        "name": "aid",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "应用名称",
                        "name": "name",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "description": "收费模式 1=扣时 2=扣点",
                        "name": "mold",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "description": "绑定模式 1=不绑定 2=绑定机器码 3=绑定IP",
                        "name": "bin",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "description": "登录模式 1=免费 2=单点登录 3=多点登录",
                        "name": "login",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "description": "远程数据",
                        "name": "msg",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "description": "1=可用 2=不可用",
                        "name": "state",
                        "in": "formData"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"code\":200,\"msg\":\"修改成功\"}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/apply/get": {
            "post": {
                "description": "数据已加密处理",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "应用管理"
                ],
                "summary": "获取应用信息[数据已加密]",
                "parameters": [
                    {
                        "type": "string",
                        "description": "应用ID",
                        "name": "aid",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"code\":200,\"msg\":\"查询成功\",\"data\":{\"aid\":1002,\"name\":\"QQ注册1\",\"mold\":2,\"login\":1,\"bin\":3,\"msg\":\"关键参数2333\",\"state\":2,\"time\":1625000360}}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/apply/getAll": {
            "get": {
                "description": "详细描述",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "应用管理"
                ],
                "summary": "获取应用列表",
                "parameters": [
                    {
                        "type": "string",
                        "description": "登录返回的token",
                        "name": "token",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "页数",
                        "name": "page",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "获取的数量",
                        "name": "limit",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "应用ID",
                        "name": "aid",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "应用状态1=可用 2=不可用",
                        "name": "state",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "应用名称",
                        "name": "name",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"code\":200,\"msg\":\"查询成功\",\"count\":2,\"data\":[{\"aid\":1001,\"name\":\"JD注册\",\"mold\":1,\"login\":3,\"bin\":2,\"msg\":\"关键参数\",\"state\":1,\"time\":1625000292}]}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/key/cread": {
            "post": {
                "description": "详细描述",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "卡密管理"
                ],
                "summary": "生成卡密",
                "parameters": [
                    {
                        "type": "string",
                        "description": "登录返回的token",
                        "name": "token",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "生成数量",
                        "name": "num",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "收费模式 1=扣时 2=扣点",
                        "name": "mold",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "点数",
                        "name": "point",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "小时数",
                        "name": "day",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "应用ID",
                        "name": "aid",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"code\":200,\"msg\":\"生成成功\",\"data\":[{\"kid\":2001,\"cdk\":\"acff954d-9160-4cc3-96c3-e9284543a68b\",\"mold\":1,\"point\":100,\"day\":24,\"uid\":0,\"aid\":1002,\"cread_time\":1625006496,\"use_time\":0,\"state\":1}]}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/key/edit": {
            "post": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "卡密管理"
                ],
                "summary": "编辑卡密信息",
                "parameters": [
                    {
                        "type": "string",
                        "description": "登录返回的token",
                        "name": "token",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "卡密ID",
                        "name": "kid",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "1=可用 2=不可用",
                        "name": "state",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"code\":200,\"msg\":\"修改成功\"}}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/key/getAll": {
            "get": {
                "description": "详细描述",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "卡密管理"
                ],
                "summary": "获取卡密列表",
                "parameters": [
                    {
                        "type": "string",
                        "description": "登录返回的token",
                        "name": "token",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "页数",
                        "name": "page",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "获取的数量",
                        "name": "limit",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "应用ID",
                        "name": "aid",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "应用状态1=可用 2=不可用",
                        "name": "state",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "卡密",
                        "name": "cdk",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"code\":200,\"msg\":\"查询成功\",\"count\":1004,\"data\":[{\"kid\":1005,\"cdk\":\"3c262067-384b-4ba5-80ab-14ca31cbadf5\",\"mold\":1,\"point\":100,\"day\":24,\"uid\":0,\"aid\":1002,\"cread_time\":1625006471,\"use_time\":0,\"state\":1}]}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
    }
}